Changes to building properties and config file
Enabling switch between using daysim and arcgis radiation files in
demand calculation
Refactoring
