+++ cea/default.config
@@ -139,6 +139,10 @@ use-dynamic-infiltration-calculation = false
 use-dynamic-infiltration-calculation.type = BooleanParameter
 use-dynamic-infiltration-calculation.help = set to true to use the dynamic infiltration calculation (takes longer!)
 
+use-daysim-radiation = true
+use-daysim-radiation.type = BooleanParameter
+use-daysim-radiation.help = set to false to use ArcGIS radiation as input, default is true
+
 [radiation]
 latitude =
 latitude.type = RealParameter
